fine-processing/requirements.txt
decorator==4.1.2
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
olefile==0.44
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
